Raw decision. Without testing
The CreateCandidate file was decomposed, the amount of code was reduced by dividing it into modules.
