+++ plugins/recruit-resources/src/components/candidate/CandidateChannels.svelte
@@ -1,0 +1,31 @@
+<!--
+// Copyright © 2020 Anticrm Platform Contributors.
+//
+// Licensed under the Eclipse Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License. You may
+// obtain a copy of the License at https://www.eclipse.org/legal/epl-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//
+// See the License for the specific language governing permissions and
+// limitations under the License.
+-->
+<script lang="ts">
+  import { ChannelsDropdown } from '@hcengineering/contact-resources'
+  import { AttachedData, Channel } from '@hcengineering/contact'
+
+  export let object: any
+  export let loading = false
+  export let matchedChannels: AttachedData<Channel>[] = []
+</script>
+
+<ChannelsDropdown
+  editable={!loading}
+  focusIndex={10}
+  bind:value={object.channels}
+  highlighted={matchedChannels.map((it) => it.provider)}
+  kind={'regular'}
+  size={'large'}
+/>
